Reorganized store
