+++ src/components/battle-planner/action-bar/realms-card/RealmListItem.tsx
@@ -14,8 +14,8 @@ import AddStructureIcon from '../../../../assets/add-structure-icon.png';
 import EditIcon from '../../../../assets/edit-icon.png';
 import useTintedImage from '../../../../hooks/useTintedImage';
 import { useAppDispatch } from '../../../../store';
-import { useBattleMapSelector } from '../../../../store/battleMap';
-import { realmSelectors, updateRealm } from '../../../../store/battleMap/realmsSlice';
+import { useBattleMapSelector } from '../../../../store/battle-planner/battle-map';
+import { realmSelectors, updateRealm } from '../../../../store/battle-planner/battle-map/realmsSlice';
 import { DEFAULT_REALM_COLORS } from '../../../../utils/gameData';
 import ColoredAvatar from '../../../common/ColoredAvatar';
 import CustomIcon from '../../../common/CustomIcon';
