+++ src/components/battle-planner/mouse-position-display/MousePositionDisplay.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from '@mui/material';
 
-import { useBattleMapSelector } from '../../../store/battleMap';
+import { useBattleMapSelector } from '../../../store/battle-planner/battle-map';
 
 function MousePositionDisplay() {
   const currentPosition = useBattleMapSelector((state) => state.mapInfo.currentMouseHover);
